ai/requirements.txt
flask==3.0.3
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
Pillow>=10.0.0
numpy>=1.24.0
gunicorn==21.2.0
flask-cors
huggingface_hub